uterm: move the uterm static library to its own directory
