+++ src/uterm/vt_fake.c
@@ -50,8 +50,8 @@
 
 #include "shl/log.h"
 #include "shl/misc.h"
-#include "uterm_vt.h"
-#include "uterm_vt_internal.h"
+#include "vt.h"
+#include "vt_internal.h"
 
 static int fake_activate(struct uterm_vt *vt)
 {
